Extract cmd subcommand to cmd_cmd.go
Move cmd flags, command definition, and run function to separate file.
